requirements_stable.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
